Added the image controller to upload images instead of the "/uploadImage" endpoints in the UserController and ArticleController
