+++ src/main/java/com/online/students/service/API/Articles/ArticleService.java
@@ -1,8 +1,5 @@
 package com.online.students.service.API.Articles;
 
-import org.springframework.web.multipart.MultipartFile;
-
-import java.io.IOException;
 import java.util.List;
 
 public interface ArticleService {
@@ -12,8 +9,6 @@ public interface ArticleService {
     ArticleEntity getById(Long id);
 
     ArticleEntity create(ArticleDTO articleDTO);
-
-    String uploadImage(MultipartFile multipartFile) throws IOException;
 
     boolean delete(Long id);
 }
