+++ src/main/java/com/online/students/service/API/Articles/ArticleServiceImpl.java
@@ -2,13 +2,10 @@ package com.online.students.service.API.Articles;
 
 import com.online.students.service.API.ArticlesCategories.ArticlesCategoryEntity;
 import com.online.students.service.API.ArticlesCategories.ArticlesCategoryService;
-import com.online.students.service.API.ImageUploading.ImageUploadingService;
 import com.online.students.service.API.Users.UserEntity;
 import com.online.students.service.API.Users.UserService;
 import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
 
-import java.io.IOException;
 import java.util.List;
 
 @Service
@@ -20,14 +17,10 @@ public class ArticleServiceImpl implements ArticleService {
 
     private final ArticlesCategoryService articlesCategoryService;
 
-    private final ImageUploadingService imageUploadingService;
-
-    public ArticleServiceImpl(ArticleRepository articleRepository, UserService userService, ArticlesCategoryService articlesCategoryService,
-                              ImageUploadingService imageUploadingService) {
+    public ArticleServiceImpl(ArticleRepository articleRepository, UserService userService, ArticlesCategoryService articlesCategoryService) {
         this.articleRepository = articleRepository;
         this.userService = userService;
         this.articlesCategoryService = articlesCategoryService;
-        this.imageUploadingService = imageUploadingService;
     }
 
     @Override
@@ -47,11 +40,6 @@ public class ArticleServiceImpl implements ArticleService {
 
         ArticleEntity article = new ArticleEntity(articleDTO.title(), articleDTO.description(), articleDTO.image(), creator, articlesCategory);
         return articleRepository.save(article);
-    }
-
-    @Override
-    public String uploadImage(MultipartFile image) throws IOException {
-        return imageUploadingService.upload(image);
     }
 
     @Override
