+++ src/main/java/com/online/students/service/API/Users/UserService.java
@@ -1,8 +1,5 @@
 package com.online.students.service.API.Users;
 
-import org.springframework.web.multipart.MultipartFile;
-
-import java.io.IOException;
 import java.util.List;
 
 public interface UserService {
@@ -13,7 +10,7 @@ public interface UserService {
 
     UserEntity create(UserDTO userDTO);
 
-    UserEntity uploadImage(Long userId, MultipartFile multipartFile) throws IOException;
+    UserEntity editAvatar(Long userId, String filename);
 
     boolean delete(Long id);
 }
